backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.9
